requirements.txt
django
pillow
djangorestframework
markdown
drf-writable-nested
mysqlclient
django-widget-tweaks